Added Vulkan Debug layer & destroy render surface
Wrapped up the implementation of the validation layer and it's error callback.
Found the render surface Destroy() was never called so made sure it did with a short term fix.

also moved some member variables and functions around which were added yesterday in the VulkanRenderSurface.h to be a tad cleaner